Refactor: query execution.
