--- app/Business/Interfaces/IDatabase.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App\Business\Interfaces;
-
-interface IDatabase 
-{
-    public function getDatabaseIdentifier();
-    public function getDatabaseDrive();
-}
