--- app/Business/Interfaces/IQueryExecutor.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Business\Interfaces;
-
-interface IQueryExecutor 
-{
-	public function execute();
-}
