+++ app/Business/Query/Executors/InsertExecutor.php
@@ -1,23 +1,30 @@
 <?php
 
-namespace App\Business\Query\Executor;
+namespace App\Business\Query\Executors;
 
-use App\Business\Session\ExecutorConstants;
+use App\Business\Query\ExecutorConstants;
+use App\Business\Query\QueryExecutor;
+use Exception;
 
 /**
- * A Executor for insert queries.
+ * Executor for Queries of insert type.
  *
- * @package App\Business\Query\Executor
+ * @package App\Business\Query\Executors
  */
 class InsertExecutor extends QueryExecutor
 {
     /**
-     * The query type keyword.
+     * Gets the keyword of the type of the Query of the Executor.
+     *
+     * @return string
      */
-    public $queryTypeKeyword = ExecutorConstants::KEYWORD_INSERT;
+    public function getTypeKeyword()
+    {
+        return ExecutorConstants::KEYWORD_INSERT;
+    }
 
     /**
-     * Execute the query.
+     * Executes Query and sets the response of execution.
      */
     public function execute()
     {
@@ -25,14 +32,14 @@ class InsertExecutor extends QueryExecutor
         {
             $result = $this->connection->insert($this->query);
 
-            $responseMessage = "Query executed successfully! $result affected rows."; // TODO: Move to const
+            $responseMessage = "Query executed successfully! $result affected rows.";
 
             $successResponse = $this->getSuccessResponse();
             $successResponse->setResponseMessage($responseMessage);
 
             $this->response = $successResponse->getJson();
         }
-        catch (\Exception $exception)
+        catch (Exception $exception)
         {
             $errorResponse = $this->getErrorResponse();
             $errorResponse->setResponseException($exception);
